[Spec Kit] Phase 8 polish: lint clean, mypy strict, hardcoded-literal scanner
Tests (T069, T074):
- tests/integration/test_determinism.py: full pipeline byte-equal across
  two runs (SC-007); confirms 20+ cross-view consistency checks pass.
- tests/unit/test_no_hardcoded_literals.py: C1 remediation — AST-walks
  every src/engine/*.py and fails on any non-allowlisted numeric literal.
  Whitelist: {0, 1, -1, 2, 52, 100, 1000, 10000} + Field(...) args +
  Literal/Annotated subscripts + lines tagged `# allow: literal`.
  Closes the Principle II enforcement gap flagged in /speckit-analyze.

Type discipline (T070):
- pyproject.toml: added pandas-stubs and types-PyYAML to dev deps;
  ignored RUF001/002/003 (Unicode math symbols ×, −, – are deliberate
  in user-facing strings and docstrings, not typos).
- src/engine/briefing.py: extracted _coerce() to narrow getattr's Any
  to int|str|None; PartnerStatus now imported for the cancel-rate
  helper's annotation.
- src/engine/classification.py: extracted _gap_above_threshold() so
  mypy can prove the short-circuit through the None check.
- src/engine/ab_test.py: narrowed groupby-key Hashables via str();
  WinnerLabel annotated explicitly.
- src/engine/projection.py: dict[Scenario, ...] instead of dict[str,...]
- src/engine/variance.py: guard _bps()'s None return for blended priced.

Auto-style: ruff --fix --unsafe-fixes cleaned 65 issues across 26 files
(unused imports, idiom tightening). Renamed scanner marker from
`# noqa: literal` (ruff parses noqa) to `# allow: literal` so ruff
ignores our domain-specific directive.

CLAUDE.md: added one-liner runtime + lint commands.

Verification: mypy strict (engine + config + data.schema) clean;
ruff clean; pytest 107 passed / 4 skipped in 58s. PDF tests still skip
on machines without libgobject/pango/cairo — quickstart.md flags the
brew install line.

This completes the implementation. All 74 tasks across 8 phases marked
[X] in tasks.md. The team can:
  - python -m src.cli.generate_data  → 160k bookings
  - streamlit run src/ui/app.py      → four tabs live
  - python -m src.cli.export --xlsx --html --no-llm --out exports/
                                     → finance-reader-grade artefacts

+++ src/engine/derivations.py
@@ -16,7 +16,7 @@ def payout_cents(coverage_pct: float, fare_cents: int) -> int:
 
     FR-007 canonical derivation.
     """
-    return int(round(coverage_pct * fare_cents))
+    return round(coverage_pct * fare_cents)
 
 
 def cost_of_service_cents(
@@ -29,7 +29,7 @@ def cost_of_service_cents(
     = (fee × payment_processing_pct) + servicing_cost_per_unit_cents.
     Clarification 2026-05-23 (Q3).
     """
-    processing = int(round(fee_cents * payment_processing_pct))
+    processing = round(fee_cents * payment_processing_pct)
     return processing + servicing_cost_per_unit_cents
